+++ app/forms/conversion/voluntary/create_project_form.rb
@@ -25,7 +25,7 @@ class Conversion::Voluntary::CreateProjectForm < Conversion::CreateProjectForm
       assigned_to: assigned_to,
       assigned_at: assigned_at,
       directive_academy_order: directive_academy_order,
-      sponsor_trust_required: sponsor_trust_required,
+      sponsor_trust_required: directive_academy_order,
       region: region
     )